// generated/user_pb.d.ts
export class User extends jspb.Message { 
    getId(): string;
    setId(value: string): User;
    getName(): string;
    setName(value: string): User;
    getEmail(): string;
    setEmail(value: string): User;
    getPhoneNumber(): string;
    setPhoneNumber(value: string): User;
    getRole(): User.Role;
    setRole(value: User.Role): User;
    getCreatedAt(): number;
    setCreatedAt(value: number): User;
    getUpdatedAt(): number;
    setUpdatedAt(value: number): User;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): User.AsObject;
    static toObject(includeInstance: boolean, msg: User): User.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: User, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): User;
    static deserializeBinaryFromReader(message: User, reader: jspb.BinaryReader): User;
}

export namespace User {
    export type AsObject = {
        id: string,
        name: string,
        email: string,
        phoneNumber: string,
        role: User.Role,
        createdAt: number,
        updatedAt: number,
    }

    export enum Role {
    USER = 0,
    ADMIN = 1,
    }

}